添加api

+++ biz/remote/AnalysisRes.go
@@ -7,7 +7,7 @@ type Submission struct {
 
 type PersonalRes struct {
 	Name        string                `json:"name"`
-	Penalty     int64                 `json:"penalty"`
+	Penalty     int                   `json:"penalty"`
 	SolveCnt    int                   `json:"solveCnt"`
 	Rank        int64                 `json:"rank"`
 	Submissions map[int64]*Submission `json:"submissions"`
